Move some standard date calculations into a single file, and use that file to also do the imports of Easter and lunar calendar calculations
